+++ tensorflow/core/kernels/scatter_functor_gpu.cu.h
@@ -84,7 +84,7 @@ __global__ void ScatterOpCustomKernel(T* params, const T* updates,
       // Ignore indices that are out of range.
       continue;
     }
-    int params_i = param_first_index * update_block + (i % update_block);
+    int64 params_i = param_first_index * update_block + (i % update_block);
     body(&params[params_i], ldg(updates + updates_i));
   }
 }
